+++ doctest/parts/private/context.cpp
@@ -524,7 +524,7 @@ int Context::run() {
             // random_shuffle implementation
             const auto first = &testArray[0];
             for (size_t i = testArray.size() - 1; i > 0; --i) {
-                int idxToSwap = std::rand() % (i + 1);
+                int idxToSwap = static_cast<int>(std::rand() % (i + 1));
 
                 const auto temp = first[i];
